editoast: rename models/paced_train.rs into models/train_schedule.rs
