+++ editoast/src/views/timetable/paced_train.rs
@@ -35,8 +35,8 @@ use super::AuthenticationExt;
 use crate::error::Result;
 use crate::models;
 use crate::models::Infra;
-use crate::models::paced_train::OccurrenceId;
-use crate::models::paced_train::TrainScheduleChangeset;
+use crate::models::train_schedule::OccurrenceId;
+use crate::models::train_schedule::TrainScheduleChangeset;
 use crate::models::train_schedule_set::TrainScheduleSet;
 use crate::views::AuthorizationError;
 use crate::views::infra::InfraIdQueryParam;
@@ -1498,7 +1498,7 @@ mod tests {
     use crate::models::fixtures::simple_paced_train_base;
     use crate::models::fixtures::simple_paced_train_changeset;
     use crate::models::fixtures::simple_sub_category;
-    use crate::models::paced_train::TrainScheduleChangeset;
+    use crate::models::train_schedule::TrainScheduleChangeset;
     use crate::views::path::pathfinding::PathfindingFailure;
     use crate::views::path::pathfinding::PathfindingResult;
     use crate::views::test_app::TestApp;
